[network] refactoring handshake protocol
* `transports.rs` had a lot of tests, so moving them out in their own file (and removing a TODO from @phlip9)
* move `HandshakeMsg` out of `UpgradeContext`. The rationale is that `UpgradeContext` should be clear and provide information about upgrading, where `HandshakeMsg` is an struct that is relevant only in the wire logic. We're losing an optimization that does not matter much (as we're not going to have too many connection establishment) and that we could regain elsewhere: if we want to optimize this flow we should have a `HandshakeContext` instead that: 1) can be stored in `NetworkContext` and 2) drives the handshake protocol and 3) caches the **serialized** `HandshakeMsg`.
* adding a test to the handshake protocol to make sure that chain id mismatch fails
