+++ common/bitvec/Cargo.toml
@@ -11,6 +11,8 @@ edition = "2018"
 
 [dependencies]
 libra-workspace-hack = { path = "../workspace-hack", version = "0.1.0" }
+proptest = { version = "0.10.1", default-features = true, optional = true }
+proptest-derive = { version = "0.2.0", optional = true }
 serde = { version = "1.0.114", features = ["derive"] }
 serde_bytes = "0.11.5"
 
@@ -18,3 +20,7 @@ serde_bytes = "0.11.5"
 lcs = { path = "../lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-proptest-helpers = { path = "../proptest-helpers", version = "0.1.0"}
 proptest = { version = "0.10.1", default-features = true }
+proptest-derive = { version = "0.2.0" }
+
+[features]
+fuzzing = ["proptest-derive", "proptest"]
