+++ network/Cargo.toml
@@ -18,6 +18,7 @@ hex = "0.4.2"
 once_cell = "1.4.0"
 pin-project = "0.4.23"
 proptest = { version = "0.10.1", default-features = true, optional = true }
+proptest-derive = { version = "0.2.0", optional = true }
 rand = { version = "0.7.3", features = ["small_rng"] }
 rand_core = { version = "0.5.1", optional = true }
 serde = { version = "1.0.114", default-features = false }
@@ -51,13 +52,14 @@ libra-proptest-helpers = { path = "../common/proptest-helpers", version = "0.1.0
 memsocket = { path = "../network/memsocket", version = "0.1.0" }
 network-builder = {path = "../network/builder", version = "0.1.0"}
 proptest = { version = "0.10.1", default-features = true }
+proptest-derive = { version = "0.2.0" }
 rand_core = { version = "0.5.1" }
 serial_test = "0.4.0"
 socket-bench-server = { path = "../network/socket-bench-server", version = "0.1.0" }
 
 [features]
 default = []
-fuzzing = ["proptest", "libra-proptest-helpers", "libra-types/fuzzing", "libra-network-address/fuzzing", "rand_core", "memsocket/testing", "netcore/fuzzing", "libra-crypto/fuzzing"]
+fuzzing = ["bitvec/fuzzing", "libra-proptest-helpers", "libra-types/fuzzing", "libra-network-address/fuzzing", "libra-crypto/fuzzing", "memsocket/testing", "netcore/fuzzing", "proptest", "proptest-derive", "rand_core"]
 testing = ["libra-config/testing", "memsocket/testing", "netcore/testing"]
 
 [[bench]]
